Isolates the Blob abstraction a bit
Also creates a type system enforced difference between the camera's
coordinate system and the world coordinate system.

The transform from CameraFramedPosition to WorldPosition is unused and
has no unit tests. It's more a sketch of how we'd like to use the type
system for this thing to avoid confusion.

Should probably isolate Blob abstraction more by defining its own type,
instead of relying on cv::KeyPoint, but that didn't fit in this commit.