Reads marker positions from command line
Also sorts the detected markers to make them match the user specified
marker positions.

Does not have the tests it should have yet.

Does not handle the case of detecting fewer, or more than 6 markers,
of which 2 are red, 2 are green, and 2 are blue, yet.

Lots of tests to write.