Adds alternative sphereToEllipseWidthHeight implementation
This allows us to test the implementations by checking that they
yield the same result.

This uncovered a very subtle bug, where atan was used instead of asin,
which had introduced an extremely small error in the geometric derivation.

This is important because the geometric derivation has been an
undocumented source of assumptions in the blobDetector code.