Adds and fulfills more tricky filtering tests
The filtering is now more heavily based on position,
and less on color.

The color sorting only reduces candidate sixtuples from 6!=720 to 2!*2!*2! = 8
in the last identification step.

However, it also makes the last identification step more robust, in case
the positions of the markers happened to be close to symmetrical.

So the colors allow us to still place out the markers in a perfect
circle, and not worry about them becoming indistinguishable from each
other.