Removes topleft marker feature.
This feature registered an expected position for a marker up in the
top left corner of the image. This data was used to compensate
(linearly, so not theoretically correct) for camera tilt between images.

The feature worked ok, but was never used. It was clearly worse than
a full reference frame registration (multiple markers on background area)
would be.

Full reference frame is exactly the next feature to be implemented,
so it makes sense to remove the topleft feature now.